Provisioning: Reorganize v1beta1 tests and move helpers to common package
Reorganize tests by moving them to appropriate files and consolidate
helper functions in the common package for better code reuse.

Changes:
- Moved TestIntegrationV1Beta1RepositoryCRUD from basic_test.go to repository_test.go
- Moved spec tests (APIAvailable, OpenAPISchema) from basic_test.go to spec_test.go
- Moved v1beta1 helper functions from pkg/tests/apis/provisioning/v1beta1/helpers.go
  to pkg/tests/apis/provisioning/common/testing.go:
  - GetConnectionClientV1Beta1
  - GetRepositoryClientV1Beta1
  - ToUnstructuredConnection
  - FromUnstructuredToConnection
  - ToUnstructuredRepository
  - FromUnstructuredToRepository
  - TestGithubPrivateKeyBase64 (returns base64-encoded test RSA key)
- Updated connection_test.go to use common package helpers
- Replaced long inline base64 RSA key with common.TestGithubPrivateKeyBase64()
- Deleted pkg/tests/apis/provisioning/v1beta1/helpers.go
