+++ pkg/tests/apis/provisioning/v1beta1/connection_test.go
@@ -9,9 +9,10 @@ import (
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
 	provisioning "github.com/grafana/grafana/apps/provisioning/pkg/apis/provisioning/v0alpha1"
-	common "github.com/grafana/grafana/pkg/apimachinery/apis/common/v0alpha1"
+	commonapi "github.com/grafana/grafana/pkg/apimachinery/apis/common/v0alpha1"
 	"github.com/grafana/grafana/pkg/services/featuremgmt"
 	"github.com/grafana/grafana/pkg/tests/apis"
+	"github.com/grafana/grafana/pkg/tests/apis/provisioning/common"
 	"github.com/grafana/grafana/pkg/tests/testinfra"
 	"github.com/grafana/grafana/pkg/util/testutil"
 )
@@ -26,7 +27,7 @@ func TestIntegrationV1Beta1Connection_Create_GitHub(t *testing.T) {
 		},
 	})
 
-	client := getConnectionClient(helper)
+	client := common.GetConnectionClientV1Beta1(helper)
 	ctx := context.Background()
 	namespace := "default"
 
@@ -48,13 +49,13 @@ func TestIntegrationV1Beta1Connection_Create_GitHub(t *testing.T) {
 			},
 		},
 		Secure: provisioning.ConnectionSecure{
-			PrivateKey: common.InlineSecureValue{
-				Create: common.NewSecretValue("LS0tLS1CRUdJTiBQUklWQVRFIEtFWS0tLS0tCk1JSUV2Z0lCQURBTkJna3Foa2lHOXcwQkFRRUZBQVNDQktnd2dnU2tBZ0VBQW9JQkFRRFJvV1E5Y3I0SG9yeHoKdjV2a1I3UjVOaUhmS0g3L2xvQWg2M0dSeks2cjZCLzhqMjNjdXlucnRMYU1ZbHgwVFgwSlZCQm1HUE85dS9YRgpQV3VkKzBac05sZXAvQzREM1lBem5rYlFLeWFpd01PbDRDcnVwcmIxQVVDOEpiTnphOXViR1U3K1k2WkJDMGJwCjhxcytIR3FYeUlwVUNySytySU1rd1V4WkJFUi9sZ3ZsMGZieE9ZVWx0WmpXRW1FQjQzVkdWbFEyQitnd0d0d1UKNXBrNXFPQ2JrVUM1Zy9QOUkyd1hsejBUbW92WEJ4dW5uc1hGbWNGVHoxTFJHNXhsQVNESTJlcE5GcjBaSmFXMQp1RVlSWWdlMmJHSmdyU1pEZzJLVXZuQ1ZBM0h3cFdtNUtTSk9OMit4cnFLMXpiRmM3RjZ1RWhQUWZ4QWlDY2QyClVWbUV2R2hoQWdNQkFBRUNnZ0VBWk1kTWkvZ0dJR0NkYlozL29YTFJjZlpTdVZoY0daNnVVZUJCTklFc3RnczkKbVdzUXUyMmxQU0lmSW9WRkFrcWdZbVF1RUE5TjJuaXFzbHlSU0oxYzRUZ00wL1VlNEEvVUNyeThPV2FOUi9BWQovcjB5bHBIQ1lpM3NnNmN5U3JwZkZJVEp3NlYvYjBTdkhsaGg5OVdzYnZUamYzeTNLdEIvcWFOZUEydDRUTWNrCll3WWVpVytlQ0dSMUZyemNrcnZpVm9CR1dac084WURaMnB3M0p3dnFESWFsbWFwMXIxbjN4UndxYmcxRTVJRXIKdVZXNlRvM2loa3VhVm90Q2xhUUpDOTBJc1MyM2FzVGl3NGZLcGxoL3ZtZUpvS2svaHBHdnBJTTJ6bGdVV2FYaApoMDV3ODcvMXBpSHF1TzBFKzhhL1lraHIzdkUwdHN2UU9GaEx2TVd0TXdLQmdRRHdlWTFPK0NjZVpTSVFmbllwClZvMzNYZGgvWjFTN0FKWHBEaHM1ajZYQ0lVaXZaNTIzT2xqK1lzeUQ3YmtScitYalhqN0dQVkx4RzRXakpnbkUKUTc5azFsZWIyNDdqczJtNVRWNGRZbXhSdElKdEMwMmtMbnQ3SVBRVFh2cEtDc0lJZ3lVODNGU1gwSU9sa293cwpMWjVaZ1ZDOWZHZCszc1ZDMTBkMnUvNVZnd0tCZ1FEZktnOGZQM2pzTmRsVVFlWlpxU1QvMlZsZXc1Q2Vnait5CjBBTldzeVB5MWs0WGdzbnlQTE9jbDdsZ0dNNDNjUjArcHhyWXVrWXR6MHk1VnhrUkRYUlZQNUNRM2J0b1ZvSkUKR2Nha2VzRGxUMHNBcjk2a3IyMGVqVzVadks1ZzJzUlBxYklLTWE5STgxb1JhbXRpK3BrYjQ0MUY1MzhuM0h0MwpIcWliNlIxSlN3S0JnREZpbFIyWm90Y0FKLzNCS3QwVWRIVlBwWTJNbC84TGdMM3E4clpnaE1jWWRNZm8vSi9MCmNNbFZXdkRoR2pmQ3F2Q0Z3MWlNOFlLb2gwcFpIbnBhKzJ4bkJIanlueWF1Q3RGT1RUeTFvTThxeGZwRTd2My8KdWNZd24wOTNHdW1ueWU5Ymw0TW5NSXc0KzBBK2wyRGZRWHphTE0ydFJjZnRVZytIREpzYXdvR25Bb0dCQU1CMAozanU5bW9SallEejQ1RFk0MkE3Sm0vaE13Z0RoSlJ3SnBvZHowTEhSUGVHcXlveGM2eTFGNy9tL0NzRG5qU2dHCkp2SDNteVJRbmNOTktQSDYzM1BSY05SdVZQd0RkeTVSNkd6YTdGWVdEd0hWcWpYdWtEV0VGVUhRcGJZamxKOTcKSW04R01EdkNtczlnTHFKYXFnWlNOUGl2VDRySjY3UnNQVTdRT0pDSEFvR0JBSkRoVlFLZGZjRlJuYTh4UTY1bApyQUlOcklMQy9PL2RCcWI5Nmlqcks1RmJwTy9BcUVnWm5GaDRYYkh3KzZwcWpyYWkyMGdQSW9sdHNrZjkwSU9xCkxOZzNSdnZtNWxmeDN1Ny9MbGZKMXhieEM2dXBDdWZHZG1VMUJSZUtmSjUycHZXVFVuYUliMmFKcUZzNVV0aUIKWjIxRWlhZFhVdUxYWlk5dWt0OXdvMVRGCi0tLS0tRU5EIFBSSVZBVEUgS0VZLS0tLS0K"),
+			PrivateKey: commonapi.InlineSecureValue{
+				Create: commonapi.NewSecretValue(common.TestGithubPrivateKeyBase64()),
 			},
 		},
 	}
 
-	unstructuredObj, err := toUnstructured(connection)
+	unstructuredObj, err := common.ToUnstructuredConnection(connection)
 	require.NoError(t, err)
 
 	created, err := client.Resource.Create(ctx, unstructuredObj, metav1.CreateOptions{})
@@ -62,7 +63,7 @@ func TestIntegrationV1Beta1Connection_Create_GitHub(t *testing.T) {
 	require.NotNil(t, created)
 
 	// Verify the created object
-	createdConn, err := fromUnstructuredToConnection(created)
+	createdConn, err := common.FromUnstructuredToConnection(created)
 	require.NoError(t, err)
 	require.NotEmpty(t, createdConn.UID)
 	require.Equal(t, connection.Name, createdConn.Name)
@@ -90,7 +91,7 @@ func TestIntegrationV1Beta1Connection_Create_GitLab(t *testing.T) {
 		},
 	})
 
-	client := getConnectionClient(helper)
+	client := common.GetConnectionClientV1Beta1(helper)
 	ctx := context.Background()
 	namespace := "default"
 
@@ -111,20 +112,20 @@ func TestIntegrationV1Beta1Connection_Create_GitLab(t *testing.T) {
 			},
 		},
 		Secure: provisioning.ConnectionSecure{
-			ClientSecret: common.InlineSecureValue{
-				Create: common.NewSecretValue("gitlab-client-secret"),
+			ClientSecret: commonapi.InlineSecureValue{
+				Create: commonapi.NewSecretValue("gitlab-client-secret"),
 			},
 		},
 	}
 
-	unstructuredObj, err := toUnstructured(connection)
+	unstructuredObj, err := common.ToUnstructuredConnection(connection)
 	require.NoError(t, err)
 
 	created, err := client.Resource.Create(ctx, unstructuredObj, metav1.CreateOptions{})
 	require.NoError(t, err)
 	require.NotNil(t, created)
 
-	createdConn, err := fromUnstructuredToConnection(created)
+	createdConn, err := common.FromUnstructuredToConnection(created)
 	require.NoError(t, err)
 	require.Equal(t, provisioning.GitlabConnectionType, createdConn.Spec.Type)
 	require.Equal(t, "gitlab-client-123", createdConn.Spec.Gitlab.ClientID)
@@ -145,7 +146,7 @@ func TestIntegrationV1Beta1Connection_Create_Bitbucket(t *testing.T) {
 		},
 	})
 
-	client := getConnectionClient(helper)
+	client := common.GetConnectionClientV1Beta1(helper)
 	ctx := context.Background()
 	namespace := "default"
 
@@ -166,20 +167,20 @@ func TestIntegrationV1Beta1Connection_Create_Bitbucket(t *testing.T) {
 			},
 		},
 		Secure: provisioning.ConnectionSecure{
-			ClientSecret: common.InlineSecureValue{
-				Create: common.NewSecretValue("bitbucket-client-secret"),
+			ClientSecret: commonapi.InlineSecureValue{
+				Create: commonapi.NewSecretValue("bitbucket-client-secret"),
 			},
 		},
 	}
 
-	unstructuredObj, err := toUnstructured(connection)
+	unstructuredObj, err := common.ToUnstructuredConnection(connection)
 	require.NoError(t, err)
 
 	created, err := client.Resource.Create(ctx, unstructuredObj, metav1.CreateOptions{})
 	require.NoError(t, err)
 	require.NotNil(t, created)
 
-	createdConn, err := fromUnstructuredToConnection(created)
+	createdConn, err := common.FromUnstructuredToConnection(created)
 	require.NoError(t, err)
 	require.Equal(t, provisioning.BitbucketConnectionType, createdConn.Spec.Type)
 	require.Equal(t, "bitbucket-client-456", createdConn.Spec.Bitbucket.ClientID)
@@ -199,7 +200,7 @@ func TestIntegrationV1Beta1Connection_Get(t *testing.T) {
 		},
 	})
 
-	client := getConnectionClient(helper)
+	client := common.GetConnectionClientV1Beta1(helper)
 	ctx := context.Background()
 	namespace := "default"
 
@@ -222,13 +223,13 @@ func TestIntegrationV1Beta1Connection_Get(t *testing.T) {
 			},
 		},
 		Secure: provisioning.ConnectionSecure{
-			PrivateKey: common.InlineSecureValue{
-				Create: common.NewSecretValue("LS0tLS1CRUdJTiBQUklWQVRFIEtFWS0tLS0tCk1JSUV2Z0lCQURBTkJna3Foa2lHOXcwQkFRRUZBQVNDQktnd2dnU2tBZ0VBQW9JQkFRRFJvV1E5Y3I0SG9yeHoKdjV2a1I3UjVOaUhmS0g3L2xvQWg2M0dSeks2cjZCLzhqMjNjdXlucnRMYU1ZbHgwVFgwSlZCQm1HUE85dS9YRgpQV3VkKzBac05sZXAvQzREM1lBem5rYlFLeWFpd01PbDRDcnVwcmIxQVVDOEpiTnphOXViR1U3K1k2WkJDMGJwCjhxcytIR3FYeUlwVUNySytySU1rd1V4WkJFUi9sZ3ZsMGZieE9ZVWx0WmpXRW1FQjQzVkdWbFEyQitnd0d0d1UKNXBrNXFPQ2JrVUM1Zy9QOUkyd1hsejBUbW92WEJ4dW5uc1hGbWNGVHoxTFJHNXhsQVNESTJlcE5GcjBaSmFXMQp1RVlSWWdlMmJHSmdyU1pEZzJLVXZuQ1ZBM0h3cFdtNUtTSk9OMit4cnFLMXpiRmM3RjZ1RWhQUWZ4QWlDY2QyClVWbUV2R2hoQWdNQkFBRUNnZ0VBWk1kTWkvZ0dJR0NkYlozL29YTFJjZlpTdVZoY0daNnVVZUJCTklFc3RnczkKbVdzUXUyMmxQU0lmSW9WRkFrcWdZbVF1RUE5TjJuaXFzbHlSU0oxYzRUZ00wL1VlNEEvVUNyeThPV2FOUi9BWQovcjB5bHBIQ1lpM3NnNmN5U3JwZkZJVEp3NlYvYjBTdkhsaGg5OVdzYnZUamYzeTNLdEIvcWFOZUEydDRUTWNrCll3WWVpVytlQ0dSMUZyemNrcnZpVm9CR1dac084WURaMnB3M0p3dnFESWFsbWFwMXIxbjN4UndxYmcxRTVJRXIKdVZXNlRvM2loa3VhVm90Q2xhUUpDOTBJc1MyM2FzVGl3NGZLcGxoL3ZtZUpvS2svaHBHdnBJTTJ6bGdVV2FYaApoMDV3ODcvMXBpSHF1TzBFKzhhL1lraHIzdkUwdHN2UU9GaEx2TVd0TXdLQmdRRHdlWTFPK0NjZVpTSVFmbllwClZvMzNYZGgvWjFTN0FKWHBEaHM1ajZYQ0lVaXZaNTIzT2xqK1lzeUQ3YmtScitYalhqN0dQVkx4RzRXakpnbkUKUTc5azFsZWIyNDdqczJtNVRWNGRZbXhSdElKdEMwMmtMbnQ3SVBRVFh2cEtDc0lJZ3lVODNGU1gwSU9sa293cwpMWjVaZ1ZDOWZHZCszc1ZDMTBkMnUvNVZnd0tCZ1FEZktnOGZQM2pzTmRsVVFlWlpxU1QvMlZsZXc1Q2Vnait5CjBBTldzeVB5MWs0WGdzbnlQTE9jbDdsZ0dNNDNjUjArcHhyWXVrWXR6MHk1VnhrUkRYUlZQNUNRM2J0b1ZvSkUKR2Nha2VzRGxUMHNBcjk2a3IyMGVqVzVadks1ZzJzUlBxYklLTWE5STgxb1JhbXRpK3BrYjQ0MUY1MzhuM0h0MwpIcWliNlIxSlN3S0JnREZpbFIyWm90Y0FKLzNCS3QwVWRIVlBwWTJNbC84TGdMM3E4clpnaE1jWWRNZm8vSi9MCmNNbFZXdkRoR2pmQ3F2Q0Z3MWlNOFlLb2gwcFpIbnBhKzJ4bkJIanlueWF1Q3RGT1RUeTFvTThxeGZwRTd2My8KdWNZd24wOTNHdW1ueWU5Ymw0TW5NSXc0KzBBK2wyRGZRWHphTE0ydFJjZnRVZytIREpzYXdvR25Bb0dCQU1CMAozanU5bW9SallEejQ1RFk0MkE3Sm0vaE13Z0RoSlJ3SnBvZHowTEhSUGVHcXlveGM2eTFGNy9tL0NzRG5qU2dHCkp2SDNteVJRbmNOTktQSDYzM1BSY05SdVZQd0RkeTVSNkd6YTdGWVdEd0hWcWpYdWtEV0VGVUhRcGJZamxKOTcKSW04R01EdkNtczlnTHFKYXFnWlNOUGl2VDRySjY3UnNQVTdRT0pDSEFvR0JBSkRoVlFLZGZjRlJuYTh4UTY1bApyQUlOcklMQy9PL2RCcWI5Nmlqcks1RmJwTy9BcUVnWm5GaDRYYkh3KzZwcWpyYWkyMGdQSW9sdHNrZjkwSU9xCkxOZzNSdnZtNWxmeDN1Ny9MbGZKMXhieEM2dXBDdWZHZG1VMUJSZUtmSjUycHZXVFVuYUliMmFKcUZzNVV0aUIKWjIxRWlhZFhVdUxYWlk5dWt0OXdvMVRGCi0tLS0tRU5EIFBSSVZBVEUgS0VZLS0tLS0K"),
+			PrivateKey: commonapi.InlineSecureValue{
+				Create: commonapi.NewSecretValue(common.TestGithubPrivateKeyBase64()),
 			},
 		},
 	}
 
-	unstructuredObj, err := toUnstructured(connection)
+	unstructuredObj, err := common.ToUnstructuredConnection(connection)
 	require.NoError(t, err)
 
 	_, err = client.Resource.Create(ctx, unstructuredObj, metav1.CreateOptions{})
@@ -239,7 +240,7 @@ func TestIntegrationV1Beta1Connection_Get(t *testing.T) {
 	require.NoError(t, err)
 	require.NotNil(t, retrieved)
 
-	retrievedConn, err := fromUnstructuredToConnection(retrieved)
+	retrievedConn, err := common.FromUnstructuredToConnection(retrieved)
 	require.NoError(t, err)
 	require.Equal(t, "test-get-connection", retrievedConn.Name)
 	require.Equal(t, namespace, retrievedConn.Namespace)
@@ -260,7 +261,7 @@ func TestIntegrationV1Beta1Connection_List(t *testing.T) {
 		},
 	})
 
-	client := getConnectionClient(helper)
+	client := common.GetConnectionClientV1Beta1(helper)
 	ctx := context.Background()
 	namespace := "default"
 
@@ -283,13 +284,13 @@ func TestIntegrationV1Beta1Connection_List(t *testing.T) {
 			},
 		},
 		Secure: provisioning.ConnectionSecure{
-			PrivateKey: common.InlineSecureValue{
-				Create: common.NewSecretValue("LS0tLS1CRUdJTiBQUklWQVRFIEtFWS0tLS0tCk1JSUV2Z0lCQURBTkJna3Foa2lHOXcwQkFRRUZBQVNDQktnd2dnU2tBZ0VBQW9JQkFRRFJvV1E5Y3I0SG9yeHoKdjV2a1I3UjVOaUhmS0g3L2xvQWg2M0dSeks2cjZCLzhqMjNjdXlucnRMYU1ZbHgwVFgwSlZCQm1HUE85dS9YRgpQV3VkKzBac05sZXAvQzREM1lBem5rYlFLeWFpd01PbDRDcnVwcmIxQVVDOEpiTnphOXViR1U3K1k2WkJDMGJwCjhxcytIR3FYeUlwVUNySytySU1rd1V4WkJFUi9sZ3ZsMGZieE9ZVWx0WmpXRW1FQjQzVkdWbFEyQitnd0d0d1UKNXBrNXFPQ2JrVUM1Zy9QOUkyd1hsejBUbW92WEJ4dW5uc1hGbWNGVHoxTFJHNXhsQVNESTJlcE5GcjBaSmFXMQp1RVlSWWdlMmJHSmdyU1pEZzJLVXZuQ1ZBM0h3cFdtNUtTSk9OMit4cnFLMXpiRmM3RjZ1RWhQUWZ4QWlDY2QyClVWbUV2R2hoQWdNQkFBRUNnZ0VBWk1kTWkvZ0dJR0NkYlozL29YTFJjZlpTdVZoY0daNnVVZUJCTklFc3RnczkKbVdzUXUyMmxQU0lmSW9WRkFrcWdZbVF1RUE5TjJuaXFzbHlSU0oxYzRUZ00wL1VlNEEvVUNyeThPV2FOUi9BWQovcjB5bHBIQ1lpM3NnNmN5U3JwZkZJVEp3NlYvYjBTdkhsaGg5OVdzYnZUamYzeTNLdEIvcWFOZUEydDRUTWNrCll3WWVpVytlQ0dSMUZyemNrcnZpVm9CR1dac084WURaMnB3M0p3dnFESWFsbWFwMXIxbjN4UndxYmcxRTVJRXIKdVZXNlRvM2loa3VhVm90Q2xhUUpDOTBJc1MyM2FzVGl3NGZLcGxoL3ZtZUpvS2svaHBHdnBJTTJ6bGdVV2FYaApoMDV3ODcvMXBpSHF1TzBFKzhhL1lraHIzdkUwdHN2UU9GaEx2TVd0TXdLQmdRRHdlWTFPK0NjZVpTSVFmbllwClZvMzNYZGgvWjFTN0FKWHBEaHM1ajZYQ0lVaXZaNTIzT2xqK1lzeUQ3YmtScitYalhqN0dQVkx4RzRXakpnbkUKUTc5azFsZWIyNDdqczJtNVRWNGRZbXhSdElKdEMwMmtMbnQ3SVBRVFh2cEtDc0lJZ3lVODNGU1gwSU9sa293cwpMWjVaZ1ZDOWZHZCszc1ZDMTBkMnUvNVZnd0tCZ1FEZktnOGZQM2pzTmRsVVFlWlpxU1QvMlZsZXc1Q2Vnait5CjBBTldzeVB5MWs0WGdzbnlQTE9jbDdsZ0dNNDNjUjArcHhyWXVrWXR6MHk1VnhrUkRYUlZQNUNRM2J0b1ZvSkUKR2Nha2VzRGxUMHNBcjk2a3IyMGVqVzVadks1ZzJzUlBxYklLTWE5STgxb1JhbXRpK3BrYjQ0MUY1MzhuM0h0MwpIcWliNlIxSlN3S0JnREZpbFIyWm90Y0FKLzNCS3QwVWRIVlBwWTJNbC84TGdMM3E4clpnaE1jWWRNZm8vSi9MCmNNbFZXdkRoR2pmQ3F2Q0Z3MWlNOFlLb2gwcFpIbnBhKzJ4bkJIanlueWF1Q3RGT1RUeTFvTThxeGZwRTd2My8KdWNZd24wOTNHdW1ueWU5Ymw0TW5NSXc0KzBBK2wyRGZRWHphTE0ydFJjZnRVZytIREpzYXdvR25Bb0dCQU1CMAozanU5bW9SallEejQ1RFk0MkE3Sm0vaE13Z0RoSlJ3SnBvZHowTEhSUGVHcXlveGM2eTFGNy9tL0NzRG5qU2dHCkp2SDNteVJRbmNOTktQSDYzM1BSY05SdVZQd0RkeTVSNkd6YTdGWVdEd0hWcWpYdWtEV0VGVUhRcGJZamxKOTcKSW04R01EdkNtczlnTHFKYXFnWlNOUGl2VDRySjY3UnNQVTdRT0pDSEFvR0JBSkRoVlFLZGZjRlJuYTh4UTY1bApyQUlOcklMQy9PL2RCcWI5Nmlqcks1RmJwTy9BcUVnWm5GaDRYYkh3KzZwcWpyYWkyMGdQSW9sdHNrZjkwSU9xCkxOZzNSdnZtNWxmeDN1Ny9MbGZKMXhieEM2dXBDdWZHZG1VMUJSZUtmSjUycHZXVFVuYUliMmFKcUZzNVV0aUIKWjIxRWlhZFhVdUxYWlk5dWt0OXdvMVRGCi0tLS0tRU5EIFBSSVZBVEUgS0VZLS0tLS0K"),
+			PrivateKey: commonapi.InlineSecureValue{
+				Create: commonapi.NewSecretValue(common.TestGithubPrivateKeyBase64()),
 			},
 		},
 	}
 
-	unstructuredObj, err := toUnstructured(connection)
+	unstructuredObj, err := common.ToUnstructuredConnection(connection)
 	require.NoError(t, err)
 
 	_, err = client.Resource.Create(ctx, unstructuredObj, metav1.CreateOptions{})
@@ -326,7 +327,7 @@ func TestIntegrationV1Beta1Connection_Update(t *testing.T) {
 		},
 	})
 
-	client := getConnectionClient(helper)
+	client := common.GetConnectionClientV1Beta1(helper)
 	ctx := context.Background()
 	namespace := "default"
 
@@ -349,13 +350,13 @@ func TestIntegrationV1Beta1Connection_Update(t *testing.T) {
 			},
 		},
 		Secure: provisioning.ConnectionSecure{
-			PrivateKey: common.InlineSecureValue{
-				Create: common.NewSecretValue("LS0tLS1CRUdJTiBQUklWQVRFIEtFWS0tLS0tCk1JSUV2Z0lCQURBTkJna3Foa2lHOXcwQkFRRUZBQVNDQktnd2dnU2tBZ0VBQW9JQkFRRFJvV1E5Y3I0SG9yeHoKdjV2a1I3UjVOaUhmS0g3L2xvQWg2M0dSeks2cjZCLzhqMjNjdXlucnRMYU1ZbHgwVFgwSlZCQm1HUE85dS9YRgpQV3VkKzBac05sZXAvQzREM1lBem5rYlFLeWFpd01PbDRDcnVwcmIxQVVDOEpiTnphOXViR1U3K1k2WkJDMGJwCjhxcytIR3FYeUlwVUNySytySU1rd1V4WkJFUi9sZ3ZsMGZieE9ZVWx0WmpXRW1FQjQzVkdWbFEyQitnd0d0d1UKNXBrNXFPQ2JrVUM1Zy9QOUkyd1hsejBUbW92WEJ4dW5uc1hGbWNGVHoxTFJHNXhsQVNESTJlcE5GcjBaSmFXMQp1RVlSWWdlMmJHSmdyU1pEZzJLVXZuQ1ZBM0h3cFdtNUtTSk9OMit4cnFLMXpiRmM3RjZ1RWhQUWZ4QWlDY2QyClVWbUV2R2hoQWdNQkFBRUNnZ0VBWk1kTWkvZ0dJR0NkYlozL29YTFJjZlpTdVZoY0daNnVVZUJCTklFc3RnczkKbVdzUXUyMmxQU0lmSW9WRkFrcWdZbVF1RUE5TjJuaXFzbHlSU0oxYzRUZ00wL1VlNEEvVUNyeThPV2FOUi9BWQovcjB5bHBIQ1lpM3NnNmN5U3JwZkZJVEp3NlYvYjBTdkhsaGg5OVdzYnZUamYzeTNLdEIvcWFOZUEydDRUTWNrCll3WWVpVytlQ0dSMUZyemNrcnZpVm9CR1dac084WURaMnB3M0p3dnFESWFsbWFwMXIxbjN4UndxYmcxRTVJRXIKdVZXNlRvM2loa3VhVm90Q2xhUUpDOTBJc1MyM2FzVGl3NGZLcGxoL3ZtZUpvS2svaHBHdnBJTTJ6bGdVV2FYaApoMDV3ODcvMXBpSHF1TzBFKzhhL1lraHIzdkUwdHN2UU9GaEx2TVd0TXdLQmdRRHdlWTFPK0NjZVpTSVFmbllwClZvMzNYZGgvWjFTN0FKWHBEaHM1ajZYQ0lVaXZaNTIzT2xqK1lzeUQ3YmtScitYalhqN0dQVkx4RzRXakpnbkUKUTc5azFsZWIyNDdqczJtNVRWNGRZbXhSdElKdEMwMmtMbnQ3SVBRVFh2cEtDc0lJZ3lVODNGU1gwSU9sa293cwpMWjVaZ1ZDOWZHZCszc1ZDMTBkMnUvNVZnd0tCZ1FEZktnOGZQM2pzTmRsVVFlWlpxU1QvMlZsZXc1Q2Vnait5CjBBTldzeVB5MWs0WGdzbnlQTE9jbDdsZ0dNNDNjUjArcHhyWXVrWXR6MHk1VnhrUkRYUlZQNUNRM2J0b1ZvSkUKR2Nha2VzRGxUMHNBcjk2a3IyMGVqVzVadks1ZzJzUlBxYklLTWE5STgxb1JhbXRpK3BrYjQ0MUY1MzhuM0h0MwpIcWliNlIxSlN3S0JnREZpbFIyWm90Y0FKLzNCS3QwVWRIVlBwWTJNbC84TGdMM3E4clpnaE1jWWRNZm8vSi9MCmNNbFZXdkRoR2pmQ3F2Q0Z3MWlNOFlLb2gwcFpIbnBhKzJ4bkJIanlueWF1Q3RGT1RUeTFvTThxeGZwRTd2My8KdWNZd24wOTNHdW1ueWU5Ymw0TW5NSXc0KzBBK2wyRGZRWHphTE0ydFJjZnRVZytIREpzYXdvR25Bb0dCQU1CMAozanU5bW9SallEejQ1RFk0MkE3Sm0vaE13Z0RoSlJ3SnBvZHowTEhSUGVHcXlveGM2eTFGNy9tL0NzRG5qU2dHCkp2SDNteVJRbmNOTktQSDYzM1BSY05SdVZQd0RkeTVSNkd6YTdGWVdEd0hWcWpYdWtEV0VGVUhRcGJZamxKOTcKSW04R01EdkNtczlnTHFKYXFnWlNOUGl2VDRySjY3UnNQVTdRT0pDSEFvR0JBSkRoVlFLZGZjRlJuYTh4UTY1bApyQUlOcklMQy9PL2RCcWI5Nmlqcks1RmJwTy9BcUVnWm5GaDRYYkh3KzZwcWpyYWkyMGdQSW9sdHNrZjkwSU9xCkxOZzNSdnZtNWxmeDN1Ny9MbGZKMXhieEM2dXBDdWZHZG1VMUJSZUtmSjUycHZXVFVuYUliMmFKcUZzNVV0aUIKWjIxRWlhZFhVdUxYWlk5dWt0OXdvMVRGCi0tLS0tRU5EIFBSSVZBVEUgS0VZLS0tLS0K"),
+			PrivateKey: commonapi.InlineSecureValue{
+				Create: commonapi.NewSecretValue(common.TestGithubPrivateKeyBase64()),
 			},
 		},
 	}
 
-	unstructuredObj, err := toUnstructured(connection)
+	unstructuredObj, err := common.ToUnstructuredConnection(connection)
 	require.NoError(t, err)
 
 	_, err = client.Resource.Create(ctx, unstructuredObj, metav1.CreateOptions{})
@@ -365,27 +366,27 @@ func TestIntegrationV1Beta1Connection_Update(t *testing.T) {
 	current, err := client.Resource.Get(ctx, "test-update-connection", metav1.GetOptions{})
 	require.NoError(t, err)
 
-	currentConn, err := fromUnstructuredToConnection(current)
+	currentConn, err := common.FromUnstructuredToConnection(current)
 	require.NoError(t, err)
 
 	// Update the AppID
 	currentConn.Spec.GitHub.AppID = "999999"
 
-	unstructuredObj, err = toUnstructured(currentConn)
+	unstructuredObj, err = common.ToUnstructuredConnection(currentConn)
 	require.NoError(t, err)
 
 	updated, err := client.Resource.Update(ctx, unstructuredObj, metav1.UpdateOptions{})
 	require.NoError(t, err)
 	require.NotNil(t, updated)
 
-	updatedConn, err := fromUnstructuredToConnection(updated)
+	updatedConn, err := common.FromUnstructuredToConnection(updated)
 	require.NoError(t, err)
 	require.Equal(t, "999999", updatedConn.Spec.GitHub.AppID)
 
 	// Verify the update persisted
 	retrieved, err := client.Resource.Get(ctx, "test-update-connection", metav1.GetOptions{})
 	require.NoError(t, err)
-	retrievedConn, err := fromUnstructuredToConnection(retrieved)
+	retrievedConn, err := common.FromUnstructuredToConnection(retrieved)
 	require.NoError(t, err)
 	require.Equal(t, "999999", retrievedConn.Spec.GitHub.AppID)
 
@@ -404,7 +405,7 @@ func TestIntegrationV1Beta1Connection_Delete(t *testing.T) {
 		},
 	})
 
-	client := getConnectionClient(helper)
+	client := common.GetConnectionClientV1Beta1(helper)
 	ctx := context.Background()
 	namespace := "default"
 
@@ -427,13 +428,13 @@ func TestIntegrationV1Beta1Connection_Delete(t *testing.T) {
 			},
 		},
 		Secure: provisioning.ConnectionSecure{
-			PrivateKey: common.InlineSecureValue{
-				Create: common.NewSecretValue("LS0tLS1CRUdJTiBQUklWQVRFIEtFWS0tLS0tCk1JSUV2Z0lCQURBTkJna3Foa2lHOXcwQkFRRUZBQVNDQktnd2dnU2tBZ0VBQW9JQkFRRFJvV1E5Y3I0SG9yeHoKdjV2a1I3UjVOaUhmS0g3L2xvQWg2M0dSeks2cjZCLzhqMjNjdXlucnRMYU1ZbHgwVFgwSlZCQm1HUE85dS9YRgpQV3VkKzBac05sZXAvQzREM1lBem5rYlFLeWFpd01PbDRDcnVwcmIxQVVDOEpiTnphOXViR1U3K1k2WkJDMGJwCjhxcytIR3FYeUlwVUNySytySU1rd1V4WkJFUi9sZ3ZsMGZieE9ZVWx0WmpXRW1FQjQzVkdWbFEyQitnd0d0d1UKNXBrNXFPQ2JrVUM1Zy9QOUkyd1hsejBUbW92WEJ4dW5uc1hGbWNGVHoxTFJHNXhsQVNESTJlcE5GcjBaSmFXMQp1RVlSWWdlMmJHSmdyU1pEZzJLVXZuQ1ZBM0h3cFdtNUtTSk9OMit4cnFLMXpiRmM3RjZ1RWhQUWZ4QWlDY2QyClVWbUV2R2hoQWdNQkFBRUNnZ0VBWk1kTWkvZ0dJR0NkYlozL29YTFJjZlpTdVZoY0daNnVVZUJCTklFc3RnczkKbVdzUXUyMmxQU0lmSW9WRkFrcWdZbVF1RUE5TjJuaXFzbHlSU0oxYzRUZ00wL1VlNEEvVUNyeThPV2FOUi9BWQovcjB5bHBIQ1lpM3NnNmN5U3JwZkZJVEp3NlYvYjBTdkhsaGg5OVdzYnZUamYzeTNLdEIvcWFOZUEydDRUTWNrCll3WWVpVytlQ0dSMUZyemNrcnZpVm9CR1dac084WURaMnB3M0p3dnFESWFsbWFwMXIxbjN4UndxYmcxRTVJRXIKdVZXNlRvM2loa3VhVm90Q2xhUUpDOTBJc1MyM2FzVGl3NGZLcGxoL3ZtZUpvS2svaHBHdnBJTTJ6bGdVV2FYaApoMDV3ODcvMXBpSHF1TzBFKzhhL1lraHIzdkUwdHN2UU9GaEx2TVd0TXdLQmdRRHdlWTFPK0NjZVpTSVFmbllwClZvMzNYZGgvWjFTN0FKWHBEaHM1ajZYQ0lVaXZaNTIzT2xqK1lzeUQ3YmtScitYalhqN0dQVkx4RzRXakpnbkUKUTc5azFsZWIyNDdqczJtNVRWNGRZbXhSdElKdEMwMmtMbnQ3SVBRVFh2cEtDc0lJZ3lVODNGU1gwSU9sa293cwpMWjVaZ1ZDOWZHZCszc1ZDMTBkMnUvNVZnd0tCZ1FEZktnOGZQM2pzTmRsVVFlWlpxU1QvMlZsZXc1Q2Vnait5CjBBTldzeVB5MWs0WGdzbnlQTE9jbDdsZ0dNNDNjUjArcHhyWXVrWXR6MHk1VnhrUkRYUlZQNUNRM2J0b1ZvSkUKR2Nha2VzRGxUMHNBcjk2a3IyMGVqVzVadks1ZzJzUlBxYklLTWE5STgxb1JhbXRpK3BrYjQ0MUY1MzhuM0h0MwpIcWliNlIxSlN3S0JnREZpbFIyWm90Y0FKLzNCS3QwVWRIVlBwWTJNbC84TGdMM3E4clpnaE1jWWRNZm8vSi9MCmNNbFZXdkRoR2pmQ3F2Q0Z3MWlNOFlLb2gwcFpIbnBhKzJ4bkJIanlueWF1Q3RGT1RUeTFvTThxeGZwRTd2My8KdWNZd24wOTNHdW1ueWU5Ymw0TW5NSXc0KzBBK2wyRGZRWHphTE0ydFJjZnRVZytIREpzYXdvR25Bb0dCQU1CMAozanU5bW9SallEejQ1RFk0MkE3Sm0vaE13Z0RoSlJ3SnBvZHowTEhSUGVHcXlveGM2eTFGNy9tL0NzRG5qU2dHCkp2SDNteVJRbmNOTktQSDYzM1BSY05SdVZQd0RkeTVSNkd6YTdGWVdEd0hWcWpYdWtEV0VGVUhRcGJZamxKOTcKSW04R01EdkNtczlnTHFKYXFnWlNOUGl2VDRySjY3UnNQVTdRT0pDSEFvR0JBSkRoVlFLZGZjRlJuYTh4UTY1bApyQUlOcklMQy9PL2RCcWI5Nmlqcks1RmJwTy9BcUVnWm5GaDRYYkh3KzZwcWpyYWkyMGdQSW9sdHNrZjkwSU9xCkxOZzNSdnZtNWxmeDN1Ny9MbGZKMXhieEM2dXBDdWZHZG1VMUJSZUtmSjUycHZXVFVuYUliMmFKcUZzNVV0aUIKWjIxRWlhZFhVdUxYWlk5dWt0OXdvMVRGCi0tLS0tRU5EIFBSSVZBVEUgS0VZLS0tLS0K"),
+			PrivateKey: commonapi.InlineSecureValue{
+				Create: commonapi.NewSecretValue(common.TestGithubPrivateKeyBase64()),
 			},
 		},
 	}
 
-	unstructuredObj, err := toUnstructured(connection)
+	unstructuredObj, err := common.ToUnstructuredConnection(connection)
 	require.NoError(t, err)
 
 	_, err = client.Resource.Create(ctx, unstructuredObj, metav1.CreateOptions{})
